refactor(ui): move MSW handlers and server to separate files
